requirements.txt
# Core Dependencies
# Note: pathlib is part of Python standard library (3.4+) - no need to install
python-magic-bin>=0.4.14; sys_platform == "win32"
python-magic>=0.4.27; sys_platform != "win32"
Pillow>=10.0.0
tqdm>=4.66.0

# Optional ML/AI Libraries
scikit-learn>=1.3.0
nltk>=3.8.0

# Configuration
pyyaml>=6.0

